+++ src/main/scala/util.scala
@@ -1,7 +1,7 @@
 /**
  * Copyright (C) 2009-2010 the original author or authors.
  */
-package org.fusesoruce.scalamd
+package org.fusesource.scalamd
 
 import java.lang.StringBuilder
 import java.util.regex.{Pattern, Matcher}