+++ src/test/scala/test.scala
@@ -1,12 +1,12 @@
 /**
  * Copyright (C) 2009-2010 the original author or authors.
  */
-package org.fusesoruce.scalamd.test
+package org.fusesource.scalamd.test
 
 import org.specs.runner.JUnit4
 import org.specs.Specification
 import java.io.File
-import org.fusesoruce.scalamd.Markdown
+import org.fusesource.scalamd.Markdown
 import org.apache.commons.io.FileUtils
 import org.apache.commons.lang.StringUtils
 import org.specs.matcher.Matcher