+++ changelog.md
@@ -1,7 +1,12 @@
 Scala Markdown
 ===============================
 
-[Scala Markdown 1.0](http://scalamd.fusesource.org/maven/1.0), September 9, 2010
+[Scala Markdown 1.1](http://scalamd.fusesource.org/maven/1.1), September 14, 2010
+----
+
+* Fixed package name typo
+
+[Scala Markdown 1.0](http://scalamd.fusesource.org/maven/1.0), September 13, 2010
 ----
 
 * Initial Release